Add functionality to MainActivity
