+++ app/src/main/java/com/example/stardewvalleypockettracker/ChecklistActivity.java
@@ -3,13 +3,13 @@ package com.example.stardewvalleypockettracker;
 import android.os.Bundle;
 import android.view.View;
 
-public class EventsActivity extends BaseActivity {
+public class ChecklistActivity extends BaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.toolbar);
-        setTitle(R.string.activity_events);
+        setTitle(R.string.activity_checklists);
         setSettingsVisibility(View.INVISIBLE);
     }
 }
