+++ app/src/main/java/com/example/stardewvalleypockettracker/SettingsActivity.java
@@ -1,7 +1,5 @@
 package com.example.stardewvalleypockettracker;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.os.Bundle;
 import android.view.View;
 
@@ -10,7 +8,7 @@ public class SettingsActivity extends BaseActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.toolbar);
         setTitle(R.string.activity_settings);
         setSettingsVisibility(View.INVISIBLE);
     }
