+++ app/src/main/res/values/strings.xml
@@ -3,5 +3,5 @@
     <string name="activity_main">Home</string>
     <string name="activity_settings">Settings</string>
     <string name="activity_events">Events</string>
-    <string name="logo">logo</string>
+    <string name="activity_checklists">Checklists</string>
 </resources>
